2. CaeserCipher.py
Write a program for encrypting and decrypting text using a Caesar cipher, which is a simple type of substitution cipher. In a Caesar cipher, each letter of text is replaced with the letter that is K positions later in the alphabet, where the parameter K is an integer that serves as an encryption key. When calculating the replacement letter, if we pass the end of the alphabet (‘Z’) then we wrap back to the beginning (‘A’). Similarly, if K is negative and we pass the beginning of the alphabet, then we wrap back to the end. A text that was encrypted using key K can be decrypted using the key (- K).

The first input line contains a positive or negative integer K indicating the number of alphabetical positions to shift: this is the encryption key. The second input line is the text to be encrypted.

Note that it is possible that K > 25 or K < -25! If K is large, then when encrypting or decrypting a letter you might wrap around more than once. If K = 26 (or any positive or negative multiple of 26) then each input letter will map to the same output letter.

While encrypting or decrypting, any lowercase letters should become uppercase and any nonalphabetical characters should remain unchanged.

Sample input:

3
Hello world!

Output:

KHOOR ZRUOG!

Sample input #2:

53
Hello world!

Output:

IFMMP XPSME!

Sample input #3:

-52
Hello world!

Output:

HELLO WORLD!

Sample input #4:

2
the quick brown fox jumps over the lazy dog

Output:

VJG SWKEM DTQYP HQZ LWORU QXGT VJG NCBA FQI


Hints:

Transform each letter into a number from 0 to 25. The ord() and chr() functions will be helpful.