3. Matrix_symetry.py
The standard input consists of N rows, each containing N integers separated by spaces. Assume that N is positive, less than 50. Thus determines an input square matrix of size N x N. Determine whether the given matrix is symmetric along the main diagonal, whether it is symmetric along the minor diagonal, and whether it is symmetric along the vertical axis passing through the center of the matrix. The program outputs three integers separated by spaces to the standard output, which after 1 if the symmetry is valid, 0 if the symmetry 0 does not hold.

Example input:

1 2 3  
2 1 4  
3 4 1  

Relevant output:

1 0 0