4. Magic_Square.py
A magic square is a square matrix of numbers in which every row, column and diagonal has the same sum.

For example, here is a magic square of size 3 x 3:

6 7 2
1 5 9
8 3 4
Every row, column, and diagonal of this square has the sum 15.

Write a program that reads a square matrix of integers from standard input. The matrix will have size N x N for some N >= 2. Each input line will contain N integers, separated by one or more spaces.

In the input matrix, exactly one number will be 0. Determine whether it is possible to replace the 0 with any integer to make the square magic. If so, write out the resulting magic square. If not, write "Can't be magic".

Sample input:

6 7 2
1 5 9
8 3 0

Output:

6 7 2
1 5 9
8 3 4

Sample input #2:

8 4 3
6 0 7
1 9 5

Output:

Can't be magic

Sample input #3:

9 2 7
0 6 4
5 10 3

Output:

Can't be magic

Sample input #4:

20   7   6  17
 0  14  15  12
 9  10  11  16
 8  19  18   5

Output:

Can't be magic

Sample input #5:

9 0 7
4 6 8
5 10 3

Output:

9 2 7 
4 6 8 
5 10 3

Sample input #6:

109 0 107
104 106 108
105 110 103

Output:

109 102 107 
104 106 108 
105 110 103


Hints:

First fill in the 0 with an appropriate value, then test whether the resulting square is magic.

You might want to write a helper function that tests whether a square is magic.

The built-in function sum(), which returns the sum of a list, may be useful.