Create anagrams.py
On the input is a number N, then N rows with entries (words in the dictionary), then the number M and finally M lines with queries (words).

For each query, find among the entries all of its anagrams, that is, the words that are composed of of the same letters in the same numbers, but perhaps in a different order (each word is its own anagram). List the anagrams for each query on a separate line in alphabetical (lexicographic) order, separated by spaces. If the query is not found among the entries no anagram is found for the query, print a blank line.

Example input:

3
delta
datel
athlete
2
tadle
xyzzy
Example output:

datel delta
(blank line)